requirements.txt
streamlit
nltk
inflect
transformers
tensorflow
scikit-learn
numpy
pandas
joblib
Pillow
torch
